[FABLG-1] Pass context to health checks
Pass a context on to the healthhecks.
The context has a configurable timeout
and it's parent (which is passed in via
the http.Request) was a cancel.

If the parent's cancel is triggered, simply
return as this means that the HTTP client
terminated its connection.

If the tiemout is triggered, return
HTTP status of request timeout

Change-Id: I52dc37d37b7ae7908fb7fdd98e41600d5a25bf2a
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>